requirements.txt
streamlit
agno
python-dotenv
plotly
openai